Restructuring the project directories.
