+++ src/libs/index.js
@@ -1,0 +1,11 @@
+/**
+ * Required Modules.
+ */
+const parser = require('./parser.lib');
+
+/**
+ * Export all available modules.
+ */
+module.exports = {
+  ...parser,
+};
